refactor: extract prompt helpers from preflight to prompts.go
